\name{effdisR-package}
\alias{effdisR-package}
\alias{effdisR}
\docType{package}
\title{
\packageTitle{effdisR}
}
\description{
\packageDescription{effdisR}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{effdisR}
\packageIndices{effdisR}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{effdisR}

Maintainer: \packageMaintainer{effdisR}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
 
}
