\name{japan-effdis-estimates}
\alias{japan-effdis-estimates}
\docType{data}
\title{
Japanese long-line Task II fishing effort 1970-2009.
}
\description{
An example of Japanese long-line fishing effort calculated using the effdisR library.
}
\usage{data("japan-effdis-estimates")}
\format{
  The format is:
 chr "japan-effdis-estimates"
}
\details{
This the result of fitting with the 2-stage GAM process and then raising with Task I data.
}
\source{
http://rscloud.iccat.int/effdis/#
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(japan-effdis-estimates)
## maybe str(japan-effdis-estimates) ; plot(japan-effdis-estimates) ...
}
\keyword{datasets}
