\name{seas}
\alias{seas}
\docType{data}
\title{
Boundaries for the world seas and oceans
}
\description{
The shape file dataset (World_Seas) used by this function represents the boundaries of the major oceans and seas of the world. The source for the boundaries is the publication 'Limits of Oceans & Seas, Special Publication No. 23' published by the IHO in 1953. The dataset was composed by the Flanders Marine Data and Information Centre and are available for download here, http://www.marineregions.org/ . 
}
\usage{data("seas")}
\format{
  The format is:
 chr "seas"
}
\details{
This is a standard shapefile.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(seas)
## maybe str(seas) ; plot(seas) ...
}
\keyword{datasets}
